refactoring thunk code to sagas
