--- src/server/sagas/database.js
@@ -1,1 +1,0 @@
-// all communication between sagas and database is done in these workers
