+++ src/server/sagas/exports.js
@@ -1,2 +1,3 @@
 export * from './auth'
-export * from './database'
+export * from './friends'
+export * from './viewer'
