+++ src/server/sagas/index.js
@@ -1,10 +1,11 @@
-import { watchRemote } from 'redux-saga-sc'
+import { watchRemote, watchEmits } from 'redux-saga-sc'
 
 import {
   watchUserCreate,
   watchAuthenticateRequest,
   watchCheckEmailExistRequest,
   watchViewerLoggedIn,
+  watchFriends,
 } from './exports'
 
 export default function *sagas(socket, database, redis) {
@@ -13,6 +14,8 @@ export default function *sagas(socket, database, redis) {
     watchAuthenticateRequest(socket, database, redis),
     watchCheckEmailExistRequest(socket, database, redis),
     watchViewerLoggedIn(socket, database, redis),
-    watchRemote(socket)
+    watchFriends(socket, database, redis),
+    watchRemote(socket),
+    watchEmits(socket),
   ]
 }
