+++ src/server/selectors/viewer.js
@@ -1,0 +1,2 @@
+export const getFriendIds = state => state.getIn(['viewer', 'friendIds'])
+export const getInvites = state => state.getIn(['viewer', 'invites'])
